Move Auth to components
