+++ src/server/components/auth/__tests__/authenticated-user.test.ts
