+++ src/server/components/auth/authenticated-user.ts
